dbeaver/pro#9518: refactor pids extraction
